Add pidfile_chown to pidfile.c
This enables a change in ownership of the PID file without exposing the path of the PID file.  This is primarily used when dropping priveleges as raised priveleges are sometimes necessary to create the PID file, but the current user may not have permission to delete the PID file after dropping privelges.